Add rake task to find and delete all duplicate content
The existing `delete:duplicates` task requires a list of content IDs or
paths when deleting duplicate entries. This is quite unwieldy when
deleting large numbers of entries, such as the ones caused by the recent
rummager `_type` bug.

To speed up the duplicate cleanup process, this new rake task finds the
duplicate content before deleting it.