Ensure the metafields are preserved when amending a document
When amending a document, which happens when the publishing API sends an
update, copy the `_type` and `_id` fields directly to the new
Elasticsearch item. Currently, they are copied from the nested `_source`
data, which has caused problems when the main `_type` and the `_source`
`_type` are out of sync.