Allow duplicate deleter to handle contacts with different IDs
The duplicate deleter was skipping over duplicate HMRC contacts because
their `_id` fields were different, even though their content IDs and
links were the same. This is because the `edition` duplicate had been
created with the wrong `_id`: contact IDs do not have a leading `/`.

Add this special case to the duplicate deleter so that it deletes the
`edition` version if it also has the wrong `_id`.