Ensure duplicate deleter uses the correct ID for contacts
The duplicate deleter assumed that duplicated content always had
the same `_id` field. This isn't true for HMRC contacts, where
the link was the same but the `_id` of the bad version had a
leading `/`. This meant that the duplicate deleter sometimes
tried to delete non-existent items with the correct `_type` but
the wrong `_id`.

Fix this by looking up the correct `_id` to delete rather than
picking one of the duplicate `_id`s arbitrarily.