Move default `_type` assignment to POST request
Add more validation of search content `_type`, and move the fallback to
"edition" to the /documents POST endpoint, which is the only place which
is permitted to create documents without supplying a specific type.

This should help us avoid situations like the recent corrupted search
data, in which the `_type` was not being specified properly and was
falling back to "edition" even for specialist document types.