Add script for comparing Elasticsearch dumps
Add a `compare_es_dumps` script which compares the output from two runs
of `es_dump` and determines whether any critical data has changed.

This was written to test changes to the overnight reindexing job.

The script is quite basic and has not been tested with a wide variety of
input, but it seemed worth adding it to the repo in case it's useful in
future.