Don't ignore link field when comparing Elasticsearch dumps
Stop ignoring the `primary_publishing_organisation` field when comparing
items in two ES dumps. This field was previously ignored because it had
just been added, and so it would have overwhelmed the output of the
`compare_es_dumps` script because its value would have changed for every
item in search.