Add a debug option to use dfs_query_then_fetch option
This is a temporary test and shouldn't be released to production in its
current form.

We noticed several differences to search results for the same queries when
- changing the count parameter in https://infinite-escarpment-22010.herokuapp.com/result?search_term=water&count=10&which_test=search_match_length
- comparing results accross environments

We think this is because local IDF can differ from global IDF in multi
node clusters. See
https://www.elastic.co/guide/en/elasticsearch/guide/current/relevance-is-broken.html

When we are evaluating changes to search, we don't want to falsely
attribute these differences to the things we changed.

We're going to see if using dfs_query_then_fetch gives us a more stable
comparison if so, this will give us more confidence in using the healthcheck.