Convert government index boost into format boosts
Remove the search query boost based on index, and replace it with boosts
to the indivual search formats.

This produces a much simpler Elasticsearch query, because it no longer
has to generate one weighted query for the `government` index and a
separate unweighted query for everything else.

This should have almost no effect on search results. The only difference
will be in the few formats which are shared between `government` and
`mainstream`: `edition`, `finder` and `inside-government-link`. However,
this just affects 23 mainstream documents and so the overall change will
be very small.