Add rake task to back-populate metadata
This is intended as a one off script to add `content_store_document_type` to
existing documents. Most newly published documents should already be sending
this to rummager (but we need to check this).

This field is really useful for debugging and analytics. If it's populated
reliably, it's more useful than `format` for format boosting, because `format`
contains stuff like `publication`, which covers lots of very different document
types.

Also add in `publishing_app` and `rendering_app`, and `content_id` (if missing).

There are about 53,000 documents with missing content store document types:

https://www.gov.uk/api/search.json?filter_content_store_document_type=_MISSING&count=0