Add A/B test of format boosts
In the B variant, remove the down-weighting of content with formats used
in the government index. Replace it with an up-weighting of guidance
content.

The government index down-weighting was originally added when there was
a clear split between specialist content int the government index and
mainstream content in the mainstream index. This distinction is no
longer as clear because new specialist content has been added to the
mainstream index.

The B variant uses the `navigation_document_supertype` field to boost
guidance content. This should push truly mainstream content (as opposed
to content in the `mainstream` index) higher in search results, helping
mainstream users find content faster.

A boost of 2.5 was chosen because it is the inverse of the 0.4 boost
that is applied to government index formats in the A variant.