Down-weight FOI releases in search results
Page view data shows that users click on pages of this document type
disproportionately compared to other methods of navigation. This
suggests that they are generally searching for other content but click
on FOI releases because the search results are not clear enough.

This may be because it's not always clear from a page title whether the
page is about an FOI release or about guidance for a particular service.

Decreasing the search ranking for these pages should make it easier to
find the pages that they are normally searching for while still making
it possible to find the FOI releases themselves.