Extract DuplicateDeleter search wiring
At the moment you need to wire up a bunch of classes to run searches
from within rake tasks and standalone scripts.

Extract all this code into the SearchConfig class.

When the app runs, sinatra sets a global SearchConfig instance.
This pulls in all the configuration from files.

When the tests run, we normally stub the global `search_config` object set on
the app.  This change allows the search config to be dependency injected, so we
don't have to depend on the whole app in scripts and rake tasks.