Don't rely on sinatra global objects from search code
Currently, various parts of the search code call out to the global SearchConfig
object set by Sinatra when running as a rack application.

We want to use lib/search code in other contexts, without pulling the entire app
with it.

As a first step, use dependency injection to pass everything we need in from
calling code.

In the future, we'd like to remove some of the layers of indirection here, so that
it's easier to understand how the query is built and executed.