Add rake task to bulk delete items from an index
Around 130 DFID reports appear twice in search results: one result is a
direct link to the page, and one is to a redirect to the correct page.

The duplicates appear because the base paths of these reports were
updated in a bulk process which added the new paths to the search index
but did not remove the old ones, which have become redirects. This rake
task fixes that by deleting the old entries.

The CSV file containing the duplicate redirects can be generated from
the publishing API using the rake task created in
alphagov/publishing-api#951.