Decrease search scroll timeout
The timeout used to be set to a very high value (60 minutes) because in
Elasticsearch <0.9.12 it used to determine the total lifetime of a
scroll cursor. Scroll cursors are now refreshed on every scroll request,
so it's better to reduce the lifetime to minimise the amount of scroll
context kept in memory.