Add Rake task to export Google Analytics document data
Add a Rake task which generates a CSV containing metadata about every
document. This can be uploaded to Google Analytics where the data will
be associated with users' search events.

Uploading this data means that the page does not have to send all of
this data with every pageview, which should help performance and data
usage for users.