Add missing data fallbacks in analytics data import CSV
Fall back from `content_store_document_type` to `format` when saving
analytics data for dimension 73, and from `content_id` to `link` for the
product SKU.

The document type is more specific and more useful for analytics
purposes, but it is not populated by every publishing app. Every item in
the search index has a format and the concepts are similar, so it is
more useful than nothing.

Similarly, the content ID is currently missing from some items and the
link should be used instead.