Compare Elasticsearch indices using streams
This approach is taken to avoid the need to load
the full datset into memory.

By retrieving order data we can simple to a side by side
comparison incrementing the list position of the smallest
(string comparison) value when the record keys don't match.

This process will consume more CPU/Network/ES CPU time while
providing a stable memory usage, this should not be an issue
so long as it is run on a multiple CPU processor.