Add more test assertions about the top-level `_type` field
Documents stored in our Elasticsearch indexes currently have two fields
named `_type`: a top-level one, which is a default ES field, and a
custom one in the `_source` object.

The `_source` `_type` field is not permitted in Elasticsearch 2 so we
have made changes to make the code depend on the top-level `_type` field
instead.

This commit ensures that we are testing that the top-level `_type` field
is correct when we check the `_source` version, as a step towards moving
the latter altogether.