Remove _type and _id from the document hash
This is a blocker to upgrading Elasticsearch. When the Elasticsearch migration plugin is run, all the indexes return an error like this:

> Reserved field names ([info](https://www.elastic.co/guide/en/elasticsearch/reference/2.0/breaking_20_mapping_changes.html#migration-meta-fields))
The `_uid`, `_id`, `_type`, `_source`, `_all`, `_parent`, `_field_names`, `_routing`, `_index`, `_size`, `_timestamp`, and `_ttl` field names are reserved and can no longer be used in the document `_source`., in type: `edition`.

It's now safe to remove from the document hash, because we've changed the codebase to reference the top level
"_type" and "_id" from elasticsearch responses.

Trello: https://trello.com/c/2aAKWCwE/149-remove-reserved-id-and-type-fields-from-source-again