Remove dependencies on `_source` `_type` in test setup
Documents stored in our Elasticsearch indexes currently have two fields
named `_type`: a top-level one, which is a default ES field, and a
custom one in the `_source` object.

The `_source` `_type` field is not permitted in Elasticsearch 2 so we
have made changes such as https://github.com/alphagov/rummager/pull/791
to make the code depend on the top-level `_type` field instead.

The tests were quite inconsistent about whether they pass around the
`_type` field as part of the document hash or as a separate field. This
commit updates integration test setup to use the latter to make it more
consistent with the app code.