Distinguish `_id` and `link` fields in test setup
The `_id` field has been removed from the `_source` fields, so remove
it from the document hash in integration test setup. Tests now pass it
as a separate parameter instead.

Continue to use the `link` field as a default when tests do not specify
the `_id`.

As part of this change, make the tests more consistent when creating
documents in Elasticsearch. They were passing a mixture of string and
symbol keys to `commit_document`. These have all been updated to use
strings because that was what most tests were using already.