Create new govuk index
Also add a rake task to allow this to be done in prod.

We need to create a new index in order to replace the existing
mainstream, government and detailed indices. All queries will therefore
be done against a single index. This is the first step in the migration
process.

Mobbed with @dwhenry and @brenetic