Insert data into govuk index from Publishing API queue
Creates a new processor and presenter in the GovukIndex namespace thus keeping
it separate from the current indexing. Adds rake task to insert data. We're
only using this to test the major routing key, and therefore validations are
minimal. This points to our own version of bunny-mock, until the PR we've
opened on the original repo gets merged.

Also add the new worker to the Procfile so that it starts on deploy.