Replace test-unit-minitest dependency with test-unit
Switch dependencies to allow minitest to be upgraded in a later commit.

Rummager uses the shoulda-context which requires Test::Unit. We
previously used the test-unit-minitest gem, but it has not been updated
since 2015 and is incompatible with the lastest version of Minitest.

Use the test-unit gem instead, since it seems to be compatible with all
of our tests.