Replace minitest/colorize with minitest/pride
Switch dependencies to allow minitest to be updated in a later commit.

The minitest-colorize gem is not compatible with the latest version of
minitest, and has not been updated since 2013.

Keep the existing version of minitest (4.7.5) for now because version 5
breaks some of the unit tests and will need a separate upgrade.