Update minitest from version 4 to 5.
Update the minitest gem to a more recent version (5.6). We cannot
upgrade to the very latest version (5.10) because 5.7 causes tests to
fail.

This allows us to install some other gems for some investigation work
which we don't expect to be committed to master. It seemed worthwhile to
commit this change anyway so that rummager is using slightly more recent
versions of its dependencies.