Ensure all test aliases are set in test teardown
Most integration tests rely on test indexes and aliases being
created at the start of the test run and deleted by a separate rake
task at the end of the test run. However, the
ElasticsearchIndexGroupTest creates and deletes test indexes so it
is important that it leaves the index in a valid state afterwards.

The teardown for ElasticsearchIndexGroupTest was correctly
recreating the deleted test index, but was not creating an alias for
it. This meant that subsequent tests depended on a `mainstream_test`
index that did not exist, and would create a duplicate index with
that name rather than the more correct `mainstream_test-<date>`,
leading to intermittent test failures.

This commit fixes these failures by assigning the alias after
recreating the deleted index.