Down-weight withdrawn pages in the sitemap
Set a priority for each page in the sitemap:
- For regular, non-withdrawn pages, set the priority to 1
- For all withdrawn pages, set the priority to 0.25

The priority is a value from 0 to 1 which influences the relative
ranking of pages in search engines.

The values of 1 and 0.25 are fairly arbitrary because we have not been
able to find advice on how to pick relative priorities. The aim is to
push relevant non-withdrawn content above withdrawn content so that
users are not confused by old guidance or publications, but still keep
the withdrawn content findable in external search if it really is the
most relevant information.

https://trello.com/c/iR68Sm6R/192-downweight-withdrawn-content-in-the-sitemap