Split tests of sitemap generator and presenter
The SitemapPresenter was extracted from the SitemapGenerator in an
earlier refactoring, but the unit tests were all left on
SitemapGenerator because the logic was simple enough to test through the
generator.

This is about to change, because we plan to add more logic for
calculating page priorities based on formats. Extracting separate tests
for the presenter will make the new priority code easier to test.