Apply more boosts to page priorities in the sitemap
Internal search results are upweighted or downweighted based on several
properties of a document, such as whether it is guidance or whether it
is marked as historic (i.e. it is a publication by a previous
government).

This commit applies the same boosts to *external* search by using them
as a multiplier in the sitemap priorities. This should improve the
relative rankings of GOV.UK pages in external search results, helping
users to find the information that they're more likely to be searching
for.

https://trello.com/c/wTcs0JaG/194-add-format-weightings-to-sitemap