Ensure that sitemap priorities are always between 0 and 1
Fix a bug which would have generated priorities greater than 1 if the
configured format boosts were all less than 1.

If the maximum configured boost was, say, 0.2, then the _default_ boost
would be 1/0.2 = 5. This is outside the allowed range for the sitemap
priority.

The fix is to include the default boost of 1 when calculating the
maximum.