Increase the ElasticSearch request timeout for index creation
This fixes the intermittently failing integration tests. These were
failing because index creation sometimes takes 6 or 7 seconds, but the
request timeout was set to 5 seconds for all requests other than bulk
reloading.

Confusingly, the error seen in the integration tests was not a timeout,
but an error like this:

```
Elasticsearch::Transport::Transport::Errors::BadRequest: [400]
{"error":"IndexAlreadyExistsException[[mainstream_test-2017-07-31t09:41:05z-e5229cbf-6da5-4eb2-af27-7f9ad623ef53]
already exists]","status":400}
```

This is because the elasticsearch-transport library uses Faraday, which
uses Net::HTTP under the hood. Net::HTTP automatically retries any
idempotent requests which time out, including the PUT request to create
an index. This request _is_ idempotent, but the second request returns a
400 because the index has already been created.

https://trello.com/c/vbOLcr7b/208-fix-intermittent-indexalreadyexistsexception-error-in-tests