Centralise requires Raw extraction of requires Move all requires to a single file
This will make it easier to see the dependancies within the application, and to
load the codebase for background tasks and the console.  It also means we can
do this without having to require the sinatra application (app.rb).

In addition the application and associated webapp class have been moved out of
the base directory and into the lib/rummager folder.