Switch algorithm for calculating sitemap priorities
Use an exponential function instead of normalisation to calculate page
priorities for the sitemap.

Priorities must be numbers between 0 and 1, and should reflect the
relative priority of the page. Previously, we were calculating this by
taking the internal search boosts, normalising them against the highest
boost to get a value between 0 and 1. This had the disadvantage that the
priorties of every page (including pages without explicitly configured
boosts) would change if the maximum boost of a property was changed.

The new function uses an inverse exponential function to calculate a
priority based on the configured boosts. It has the properties:

- A page's priority is always between 0 and 1.
- The default page priority is 0.5, as specified by the sitemap
  protocol.
- A page's priority is just based on its own property boosts. The
  priority remains consistent if unrelated properties change.
- Priority ranking is consistent with boost ranking in internal search:
  if one page has a higher boost than an another, it always has a higher
  priority (within a two decimal place rounding error).