Allow sitemap format boosts to override internal search boosts
Before, the weightings applied to documents in the sitemap were the same
as those used to rank documents in internal search. The assumption was
that search results that rank highly in internal search should also rank
highly in external search because users are searching for the same
content.

But this isn't always the case: navigation pages are inappropriate
results in internal search because the user is already on the site, but
may help to orient a user coming from external search. So a boost of 0
for mainstream browse pages should be increased in the external search
ranking.