Add the ability to migrate formats from the the govuk index
This will support the migration to using the publishing API as the source of
data for search. This will allow the documents to be migarted across a format
at a time.

Please note that teh implementation used here is not compatible with ES 2.4
https://www.elastic.co/guide/en/elasticsearch/reference/2.4/query-dsl-indices-filter.html
and will need to be updated separately there.