Strip HTML tags out of indexable content
We tried several approaches when implementing this code including:

* Elasticsearch `html_strip` filter - according to the documentation
this would copy the full content to merged fields rather than the stripped
content.
* Storing the value in a separate field in ES - This would cause
inconsistencies in the scoring due to data being split across mutliple
fields resulting in documents with field not being calculated correctly
in some edge cases.
* Use of the `Rails::Html::Sanitizer` gem - This did not correctly
insert spaces when removing tags, nor did it separate block tags onto
separate lines.

Indexable content passed to elasticsearch should not have html tags in it
otherwise it can potentially effect search indexing.

The publishing api sends a few different types of payloads for
notifications it puts onto the queue therefore it is necessary to deal
with strings as well as `text/html` but to ignore `text/govspeak`.