Ensure Sidekiq workers write logs to sidekiq log file
Delete the custom logger definitions from the Sidekiq workers, which
means that logging falls back to the built-in Sidekig logging. This is
configured to write to the file in <rummager_dir>/log/sidekiq.log.

Before, the `logger` instance was the standard Ruby logger, which was
just configured to log to stdout. This meant the Sidekiq logs were not
being persisted.

One logging line had to be deleted to get this to work: the
`notify_of_failures` logging, which happens when Sidekiq retries have
been exhausted. This is because it happens in a class method, which does
not have access to the Sidekiq `logger` instance. But this event is
already logged to Airbrake, so the deleted line was redundant anyway.

Also ensure that the log output is suppressed in tests.

https://trello.com/c/zAHyGMni/240-get-logging-to-work