Limit very old publishing dates in sitemap
Some documents have been backdated with very old `public_timestamp`
dates, e.g.
https://www.gov.uk/government/publications/ukho-1914-archives-catalogue

The Google Search Console (understandably) raises an error when these
dates are used as the `lastmod` date in the sitemap.

Ideally we would use a date which corresponds to the last time the
document was updated on GOV.UK, but we do not have that information in
the search index at the moment.

As a workaround, limit the `lastmod` date to recent dates, and use the
launch of GOV.UK as a cut-off point because documents could not have
been on the site before this date. This affects around 76,000 documents.

https://trello.com/c/O2FcYxgQ/241-fix-old-dates-in-sitemap