Refactor to use a Details Presenter
A lot of the upcoming formats being moved to the govuk index make use of
elements in the details part of the payload from the publishing api.
Having a dedicated details presenter will hopefully avoid the need for
many presenters for many formats.

Also the details hash can have many things added to it. We only really
want a few specific fields to be indexed.