Attempt to describe fields in rummager
Add descriptions to some of the fields. All of them should have descriptions,
but I don't have the patience to do all them now.

There's various things that could probably be removed or simplified, but before
even thinking about this I'd like to improve the documentation quality and
start versioning the API.