Make message queue listener handle content items without base_path
The worker that listens to *.links updates never worked with content items that
don't have base_paths.

This causes the elasticsearch lookup to fail, because the term filter gets
passed null, which isn't valid. This causes messages to backup in the RabbitMQ
queue.

We're not sure why this has never caused problems before; it could be a change
in behaviour since updating elasticsearch.