Fix parameter order in test expectations
Switch the "expected" and "actual" parameters in tests of date
filtering so that failure messages are easier to interpret.

There are other tests where the parameters are the wrong way round, but
there are too many to fix them all now. This fixes the parameters in the
tests that will be used to support upcoming changes to the date filters.