Add regression test cases for date filtering
Add extra tests of the current behaviour of date parameters, such as
where the filters cut off dates which are stored as date strings or as
full date-times.

This is in preparation for changes to date filtering which will allow
dependent applications to filter search results by time as well as date.