Support date-times in date filters
Allow documents to be filtered by dates and times, as well as just by
date. The time in a filter parameter used to just be ignored.

Examples:
 - `from:2014-04-01`
 - `to:2015-12-31 14:05`
 - `from:2017-09-01 00:00:00,to:2017-09-15 23:59:59`

As well as making search filtering more flexible, this fixes a timezone
bug in finders like Announcements
(https://www.gov.uk/government/announcements) which allow users to
specify publishing dates. Before, if you entered a 'to' date which falls
during BST, the results would include anything publishing between
midnight and 1am on the following day.