Validate the 'to:' and 'from:' labels in date filtering
Reject any other labels with a 422 response rather than just ignoring
them.

We are planning to change the way date filtering works in rummager: once
we allow the date filters to include times, we will need to treat 'to'
and 'from' dates differently. This extra validation will give us more
confidence that the new filters are being applied correctly, as well as
being useful in their own right.