Don't double-count documents in aggregates
Some documents exist in both the mainstream and govuk indexes because we are currently migrating all
content to the govuk index. Search queries already look in one index or the other, based on the
document's format, which prevents duplicate search results. But aggregates (which are used to power
filters like the organisation filter on the search homepage) were not doing this, so the counts
displayed on filters were too high because documents were being counted twice.

This fix applies the same format filtering to aggregates that is already applied to search queries.

This commit also deletes some unit tests. They tested the exact Elasticsearch query, which made
them quite fragile: adding the new aggregate filter added new (and fairly complex) terms to the
query, which made it difficult to update the tests. The same functionality was being tested by
integration tests, so it seems safe to just delete the unit tests and rely on integration tests
for checking how rummager uses Elasticsearch.