Add contact format indexable content to govuk index
The contact format was not a simple mapped field. We needed to extract the titles
for each of the contacts in the `contact_groups` field. As these title
fields were at a deeper level inside `details` we could not use the string value
of the required field as we currently do with `licence` and `transaction`.

We therefore refactored indexable content into it's own presenter. This
should help us in the future when migrating other more complicated
formats.