Delete contact documents with old id format
The _id field for most content is the same as the link (including the
leading `/`), but contacts have an _id without a leading slash.

The publishing to the new govuk index will now make each contact
document have an _id with the leading `/` therefore we need to delete
all the old documents.

This task can be removed once it has been run in production after the
contact format documents have been republished to the new govuk index.