Add primary_publishing_organisation field to aggregation whitelist
To surface tagging progress in content-tagger, we use the aggregation params in
Search API to count (1) the number of documents that belong to a given primary
publishing organisation and (2) the number of these documents that have been
tagged (i.e. they have at least one taxon).

Currently, the aggregation field whitelist does not contain the
primary_publishing_organisation field we are relying on. This commit adds the
field into the whitelist which seems to satisfy our requirements.