Add A/B test of specialist sector (aka topic) boosting
We have previously run A/B tests which showed that boosting mainstream
content above specialist content helped more users to find the
information they were looking for. This new A/B test iterates on that by
boosting topics in search results, which we belive are useful navigation
pages for users to find in search.

https://trello.com/c/cmxfgx1H/391-configure-format-weighting-a-b-test-in-rummager