Handle batchs of records into the Processor
This updates the processor to take a batch of messages form the govuk message
consumer class and pass the batch to the worker. If any of the messages result
in an error from ES then the entire batch will be retried. This is to avoid the
need to implement retry functionality (backoff, limits, etc..).