Handle batch of records into the Processor
This is the first step to enabling the processing of batches from rabbit mq.
The idea is that this should increase the throughput speed when large volumes
of data are republished.

The code works by checking which of the events sent across successed, and
manually resending any that failed. Unfortunately this means the sidekiq retry
count is reset whenever more than one item in the batch is successful, however
this is considered an accetable compromise at this stage.