Add A/B test of core mainstream document_types
Modify the topic-boosting A/B test (which has not yet been launched) to
include a boost to core mainstream content.

This covers content types which we consider to be truly mainstream:
content which is useful to any member of the general public, rather than
just people in specialist roles. This is distinct from the "mainstream"
search index and the "mainstream" publisher, both of which now contain
some specialist content.

Boosting this content was one of the recommendations from the 2017 Q1
format-weighting A/B test, in which clicks on "mainstream" formats
decreased slightly even though the overall exit rate improved.

The boost value is just a starting point which we will tune using
benchmarking tests.