Add stopwords filter to synonyms analyzers
We are planning to A/B test the current synonyms approach against a new
approach which uses the `with_index_synonyms` analyzer for indexing and
the `with_search_synonyms` for queries. To make this a fair test, the
old and new analzyers should be the same apart from the synonym
handling.

This commit adds the `stop` filter to make the new analyzer consistent
with the existing analyzers (`default` for indexing and
`query_with_old_synonyms` for querying). This means that stopwords like
"the" will not be indexed in the fields used for searches with the new
synonym config, and will be stripped from users' search queries.

The `stop` filter has been added after synonyms have been applied
because synonyms may include stopwords, so we don't want to filter them
out too soon.