Prepend HMRC Manual Section ID to title
Manual section titles need to be prepended by their section ids when
being added to the elasticsearch index.

Doing this is in the common fields presenter is not ideal. As we migrate
more formats we are finding more and more instances of format specific foibles
in our payload part approach to presenters.