Configure synonym A/B test
One of the main components of a search query is the comparison with a
field which contains all the indexable content for a document (e.g.
title, description and body).

In this new A/B test, the A variant uses the `_all` field as before, but
the B variant uses the `all_searchable_text.synonym` field, which has
separate index-time and search-time analyzers that use the new synonym
definitions.

Also tidy up some previous "new" synonym code which was never used in
production.