Run all search integration tests with A and B variants
Most search queries do not change between the A and B variant of an A/B
test. For example, synonyms and format weighting do not change best bets
or quoted searches.

This commit adds a wrapper around the search integration tests to
confirm that the B version does not change these invariants to give us
more confidence that we haven't broken existing functionality.