Check first 20 results in the health check
Increase the search result count from the default (10) to 20 when
running the healthcheck.

This will make it easier to judge the healthcheck results because users
will see 20 results when the search on GOV.UK, so if a result is not
found by the healthcheck then it will not be in the first page of
results for a user.