Add a debug option to disable boosting
With this option, we can disable all format boosting, and date boosting.
You can already disable popularity boosting and best bets, and by
disabling all of those you can isolate the main text matching part of
the query, i.e. the text analysis pipeline, TF-IDF, and the boolean
queries that combine various kinds of matching.

This is very useful at the moment, because this stuff is really broken,
and it's the foundation our search query is based on - by working on
this, we can tackle some of the really weird search behaviour, where
rummager returns documents that don't look anything like the query.