Extend synonyms A/B test to other parts of the query
When we originally implemented the A/B test, we worked off the
assumption that the most important part of the query for matching
text was the "minimum should match" against the "_all" field.

I'd assumed that the "match_bigrams" part was broken and had no effect,
because no bigram tokens are created at index time.

It turns out that while this doesn't work as intended, it does have a
major contribution to the query, because the shingled_query_analyzer
preserved single-word tokens.

For example:

    $ elyzer --es 'http://dev.gov.uk:9200' --index
    govuk-2017-11-29t13:34:54z-476d0d41-e986-41bb-9595-9ebe8f0e18a1
    --analyzer shingled_query_analyzer --text "heating grants"
    TOKENIZER: standard
    {0:heating} {1:grants}
    TOKEN_FILTER: standard
    {0:heating} {1:grants}
    TOKEN_FILTER: asciifolding
    {0:heating} {1:grants}
    TOKEN_FILTER: lowercase
    {0:heating} {1:grants}
    TOKEN_FILTER: stop
    {0:heating} {1:grants}
    TOKEN_FILTER: stemmer_override
    {0:heating} {1:grants}
    TOKEN_FILTER: stemmer_english
    {0:heat}  {1:grant}
    TOKEN_FILTER: old_shingles
    {0:heat,heat grant} {1:grant}

This means that it is effectively a multi_match against 4 fields,
balanced evenly, and it *doesn't* apply synonyms.

It's very likely we could improve results by replacing this part of the
query with a simple multi match that doesn't attempt to use bigrams,
balancing the influence of the (title, description, indexable_content,
acronym, _all) fields, and then implementing shingles properly at a
later date.

For now, we should focus on cleaning up the synonyms, and that means
actually applying where they are likely to be useful. This change
ensures that all matching against these four fields uses the new
synonyms approach.