Make rubocop ignore rake tasks
Rubocop complains that our rake task use `[:param]` form of passing variables in. It is not possible to use the suggested method so just ignore all rake tasks.

The draft taxon task is a little long but will be removed once the formats have been migrated to the new index so we don't care if it's a little bit long.